move useMinimizableResourceList to new file
